# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Banka is a light-weight core banking application that powers banking operations like account creation and help in the customer to monitor his or her account and its transactions",
    "version": "1.0.0",
    "title": "MyBanka App",
    "contact": {
      "email": "princewillifeanyi1999@gmail.com"
    }
  },
  "host": "https://banka-pro-app.herokuapp.com",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/v1/transactions/{accountNumber}/credit": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Credit users account on MyBanka",
        "description": "",
        "operationId": "creditAccount",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Information account transaction made",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Info"
            }
          },
          {
            "name": "accountNumber",
            "in": "path",
            "description": "The account number of the account been whose status is being changed ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User bank account credited on MyBanka",
            "schema": {
              "$ref": "#/definitions/ApiTransactionResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BadResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/BadResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/transactions/{accountNumber}/debit": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Debit users account on MyBanka",
        "description": "",
        "operationId": "debitAccount",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Information account transaction made",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Info"
            }
          },
          {
            "name": "accountNumber",
            "in": "path",
            "description": "The account number of the account been whose status is being changed ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User bank account debited on MyBanka",
            "schema": {
              "$ref": "#/definitions/ApiTransactionResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BadResponse"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/BadResponse"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get all transactions done on MyBanka",
        "description": "",
        "operationId": "getAllTransactions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Information transactions made",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Info"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions made on MyBanka"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
  "definitions": {
...
    "Info": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "accountType": {
          "type": "string"
        }
      }
    },
...
    "ApiTransactionResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer",
          "format": "int-32"
        },
        "data": {
          "properties": {
            "id": {
              "type": "number",
              "format": "double"
            },
            "createdon": {
              "type": "string"
            },
            "transactionType": {
              "type": "string"
            },
            "accountNumber": {
              "type": "number",
              "format": "double"
            },
            "amount": {
              "type": "number",
              "format": "double"
            },
            "cashierid": {
              "type": "number",
              "format": "double"
            },
            "oldbalance": {
              "type": "number",
              "format": "double"
            },
            "newbalance": {
              "type": "number",
              "format": "double"
            },
            "userType": {
              "type": "string"
            },
            "isadmin": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "BadResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "number"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
  }
}